# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HINS Backend API",
    "version": "1.0.0",
    "description": "Contrato de API generado a partir del código fuente (no hay Swagger auto-generado en el backend). Todos los endpoints requieren `Authorization: Bearer <JWT>` salvo `/auth/register` y `/auth/login`."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" }
        },
        "required": ["access_token"]
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginDto" } } }
        },
        "responses": {
          "200": {
            "description": "Login OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "401": { "description": "Credenciales inválidas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}